+++ src/page/ConfigPageLayout/tabs/SecretsManagementTab/constants.ts
@@ -1,1 +1,2 @@
 export const SECRETS_EDIT_MODE_ADD = '__add__';
+export const SECRETS_MAX_LABELS = 10;