+++ src/page/ConfigPageLayout/tabs/SecretsManagementTab/index.ts
@@ -1,1 +1,2 @@
 export * from './SecretsManagementTab';
+export * from './types';